requirements.txt
manim>=0.18.0
edge-tts>=6.1.0
pyyaml>=6.0
